refactor: 세부 카테고리 구별 ColorSelect -> 내/외부 색상 선택 뷰로 분리 DetailModelSelect -> 엔진, 구동 방식, 바디타입 선택 뷰로 분리
